Add some basic bootstrap UI and some route refactioring. Add list instances component

+++ src/components/Recipe.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { withRouter, Redirect } from 'react-router-dom';
+import { withRouter, Redirect, Link } from 'react-router-dom';
 import { connect } from 'react-redux';
 import recipe from '../actions/recipe.js';
 import {
@@ -27,7 +27,7 @@ const mapDispatchToProps = dispatch => ({
       payload: recipe.getRecipe(recipeId, token),
   }).catch(error => {
     if (error.response.status === 404) {
-      window.location.replace('/');
+      window.location.replace('/recipes');
     }
   }),
   unloadRecipe: () => dispatch({
@@ -36,6 +36,11 @@ const mapDispatchToProps = dispatch => ({
   createRecipe: (token, name, script) => dispatch({
     type: CREATE_RECIPE,
     payload: recipe.createRecipe(token, name, script),
+  }).then(() => {
+    dispatch({
+      type: GET_RECIPES,
+      payload: recipe.getRecipes(token)
+    });
   }),
   editRecipe: (recipeId, name, script, token) => dispatch({
     type: EDIT_RECIPE,
@@ -74,6 +79,19 @@ class Recipe extends React.Component {
     this.props.getRecipes(this.props.auth.accessToken);
   }
 
+  componentWillReceiveProps(nextProps) {
+    if (nextProps.match.params.recipeId &&
+        nextProps.match.params.recipeId != this.props.match.params.recipeId) {
+      this.props.getRecipe(
+        this.props.match.params.recipeId,
+        this.props.auth.accessToken,
+      );
+    }
+    if (nextProps.recipe.deleted === true) {
+      this.props.history.push('/recipes');
+    }
+  }
+
   componentWillUnmount() {
     this.props.unloadRecipe();
   }
@@ -96,50 +114,65 @@ class Recipe extends React.Component {
     )
   }
 
-  deleteRecipe() {
-    this.props.deleteRecipe(...this.requestParams);
-  }
-
-  newRecipe() {
-    this.props.history.push('/recipe');
+  deleteRecipe(ev) {
+    ev.preventDefault();
+    if (window.confirm('Delete recipe?') === true) {
+      this.props.deleteRecipe(...this.requestParams);
+    }
   }
 
   render() {
     var recipes = this.props.recipes.map(function(recipe, i) {
-      return <div key={i}><a href={"/recipe/" + recipe.id}>{recipe.name}</a></div>;
+      return <li key={i}>
+        <div className="txt"><Link className="tip" to={'/recipe/' + recipe.id}>{recipe.name}</Link></div>
+      </li>
     });
     return(
-      <div className='container recipe'>
-        {this.props.recipe.id &&
-          <Redirect to={"/recipe/" + this.props.recipe.id}/>
-        }
-        <div className='row'><Errors errors={this.props.recipe.error} /></div>
+      <div className='container-fluid recipe'>
+        <h1>Recipes</h1>
+        <hr/>
         <div className="row">
           <div className='col-md-6'>
-            <h3>Recipes</h3>
-            {recipes}
-            <button className='btn btn-default' onClick={this.newRecipe.bind(this)} >New Recipe</button>
-          </div>
-          <div className="col-md-6">
-            <form id='recipe-form' onSubmit={this.props.recipe.id ? this.editRecipe.bind(this) : this.createRecipe.bind(this)}>
-              <Errors error={this.props.recipe.error} />
-              <div className={'form-group'}>
-                <label htmlFor='recipe_name'>Name:</label>
-                <input className={'form-control' + (this.props.recipe.errors.name ? ' has-error' : '')} onChange={this.changeName} value={this.props.recipe.name || ''} required  type='text' name='recipe_name' id='recipe_name' />
-                <Errors errors={this.props.recipe.errors.name}/>
+            <div className="widget-box">
+              <div className="widget-title">
+                <h3>Your Recipes</h3>
               </div>
-              <div  className={'form-group'}>
-                <label htmlFor='recipe_script'>Script:</label>
-                <textarea rows='20' className={'form-control' + (this.props.recipe.errors.script ? ' has-error' : '')} onChange={this.changeScript} value={this.props.recipe.script || ''} required name='recipe_script' id='recipe_script'/>
-                <Errors errors={this.props.recipe.errors.script}/>
-              </div>
-              <div className='form-group'>
-                <input className='btn btn-default' type='submit' name='submit' value='Save' id='submit' />
+              <div className="widget-content">
+                <div className="todo recipe-list">
+                  <ul>{recipes}</ul>
+                </div>
                 {this.props.recipe.id &&
-                  <button onClick={this.deleteRecipe.bind(this)} className="btn btn-danger pull-right">Delete</button>
+                  <span><Link className="btn btn-default" to="/recipes">Create New</Link></span>
                 }
               </div>
-            </form>
+            </div>
+          </div>
+          <div className="col-md-6">
+            <div className="widget-box">
+              <div className="widget-title"> <span className="icon"> <i className="icon-pencil"></i> </span>
+                <h3>New Recipe</h3>
+              </div>
+              <div className="widget-content">
+                <form id='recipe-form' onSubmit={this.props.recipe.id ? this.editRecipe.bind(this) : this.createRecipe.bind(this)}>
+                  <div className={'form-group' + (this.props.recipe.errors.name.length > 0 ? ' has-error' : '')}>
+                    <label htmlFor='recipe_name'>Name:</label>
+                    <input className='form-control' onChange={this.changeName} value={this.props.recipe.name || ''} type='text' name='recipe_name' id='recipe_name' />
+                    <Errors errors={this.props.recipe.errors.name}/>
+                  </div>
+                  <div  className={'form-group' + (this.props.recipe.errors.script.length > 0 ? ' has-error' : '')}>
+                    <label htmlFor='recipe_script'>Script:</label>
+                    <textarea rows='20' className='form-control' onChange={this.changeScript} value={this.props.recipe.script || ''} name='recipe_script' id='recipe_script'/>
+                    <Errors errors={this.props.recipe.errors.script}/>
+                  </div>
+                  <div className='form-group'>
+                    <input className='btn btn-default' type='submit' name='submit' value='Save' id='submit' />
+                    {this.props.recipe.id &&
+                      <button onClick={this.deleteRecipe.bind(this)} className="btn btn-danger pull-right">Delete</button>
+                    }
+                  </div>
+                </form>
+              </div>
+            </div>
           </div>
         </div>
       </div>
